jwt_manually_method/requirement.txt
fastapi
pydentic
uvicorn
alembic
python-dotenv
python-jose
passlib